requirements.txt
mantisconnect
zeep
python-magic
